+++ tests/codeception/e2e/app/JavaScriptCept.php
@@ -10,10 +10,18 @@ $I->amGoingTo('check javascript with a modal');
 
 $I->amOnPage('/en');
 $I->dontSee('#phd-info-panel');
+$I->dontSee('App Version');
+$I->dontSee('Project Version');
 
-$I->amGoingTo('display the info button');
+$I->login('editor', 'editor1');
+
+$I->amGoingTo('hide the info button');
 $I->pressKey("body", "h");
-$I->click('#phd-info-button a');
+$I->dontSee('#phd-info-button a');
+
+$I->amGoingTo('show the info button');
+$I->pressKey("body", "h");
+$I->click('#phd-info-button a[data-target="#phd-info-modal"]');
 $I->waitForElementVisible('#phd-info-modal',3);
 $I->seeElement('#phd-info-modal');
 $I->makeScreenshot('modal');